+++ chunkflow/operators/test_read_write.py
@@ -1,11 +1,13 @@
 import unittest
 import numpy as np
 import os
+import shutil
 
 from chunkflow.chunk import Chunk
 
 from .read_file import ReadFileOperator
 from .write_h5 import WriteH5Operator
+from .save_images import SaveImagesOperator
 
 
 class TestReadWrite(unittest.TestCase):
@@ -25,6 +27,13 @@ class TestReadWrite(unittest.TestCase):
                          chunk2.global_offset, 
                          msg='the global offset is different.')
         os.remove( file_name )
+
+        # test save images
+        output_path = '/tmp/test/'
+        save_images_operator = SaveImagesOperator(output_path)
+        save_images_operator(chunk)
+        print('remove the temporary directory.')
+        shutil.rmtree(output_path) 
 
 
 if __name__ == '__main__':